Conflicts:
	perl/openprint/Company.pm
	perl/openprint/Object.pm
	perl/openprint/Order.pm
	perl/openprint/Paper.pm
	perl/openprint/ServiceType.pm
	perl/openprint/administrator_managerial.pm
	perl/openprint/employee_accounting.pm
	perl/openprint/main_account.pm
	perl/ssi.pm
	www/public_html/email_content/additional_charges_client_notification.html
	www/public_html/email_content/additional_charges_csr_notification.html
	www/public_html/employee/account/profile.html
	www/public_html/employee/accounting/_search_results.html
	www/public_html/employee/accounting/details.html
	www/public_html/employee/accounting/search.html
	www/public_html/employee/inventory/_skid_contents.html
	www/public_html/employee/inventory/paper_label.html
	www/public_html/employee/production/project_view.html
	www/public_html/employee/project/view.html
	www/public_html/includes/login_log.html
	www/public_html/main/account/_trade_references.html
	www/public_html/main/project/reuse.html
	www/public_html/skins/PointOne Graphics Inc/css/administrator.css
	www/public_html/skins/PointOne Graphics Inc/css/common.css
	www/public_html/skins/PointOne Graphics Inc/css/email.css
	www/public_html/skins/PointOne Graphics Inc/css/main.css
	www/public_html/skins/PointOne Graphics Inc/css/press_schedule.css
	www/public_html/skins/PointOne Graphics Inc/layouts/administrator/default.html
	www/public_html/skins/PointOne Graphics Inc/layouts/default.html
	www/public_html/skins/PointOne Graphics Inc/layouts/employee/default.html

+++ perl/openprint/employee_project.pm
@@ -25,6 +25,7 @@ require openprint::PurchaseOrder_Content;
 require openprint::PaperInventory;
 require openprint::RFIDTag;
 require openprint::ScheduledJob;
+require openprint::ServiceType_Category;
 
 
 use vars qw( $r $log $dbh %variable %param %session %config );
@@ -491,6 +492,10 @@ sub send_additional_charges_notifications {
 	$info{'CompletionDate'} = Date::Format::time2str( $config{'DateTimeFormat'}, time );
 	$info{'SecureSiteURL'} = $r->dir_config('ExternalSecureSiteURL');
 	$info{'siteURL'} = $r->dir_config('ExternalSiteURL');
+	my $Project = new openprint::Project( $project_index );
+	$info{'Project'} = $Project;
+
+	openprint::project::get_header( $log, $dbh, \%info, $project_index );
 
 	my $email_template = misc::load_file( $log, $config{'SkinPath'} . '/email_template.html' );
 
@@ -523,7 +528,6 @@ sub send_additional_charges_notifications {
 			SUBJECT => 'Additional Charges required',
 			);
 	misc::send_email_with_attachment( $log, \%mail, @body );
-	my $Project = new openprint::Project( $project_index );
 	$Project->add_to_log( @session{'company_id','user_id'}, "Additional charges notification sent to : $mail{TO}." );
 
 } # End sub send_additional_charges_notifications
@@ -844,7 +848,10 @@ sub _stock_checkout {
 					$C->save();
 					#Remove any allocations
 					foreach my $PA ( openprint::PaperAllocation::find('skid_id'=>$Skid->id(),'paper_id'=>$C->paper_id(), 'docket'=>$Project->docket() ) ) {
-						$PA->delete();
+						$PA->save({'skid_ids'=>[ sets::exclude( [ $Skid->id() ], $PA->skid_ids() ) ] });
+						if ( ! $PA->Skids() ) {
+							$PA->delete();
+						} # end if
 					} # end foreach
 				} # end foreach C
 			} else {
